money refactoring

+++ rails/invest/config/initializers/custom_methods.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require Rails.root.join('lib', 'core_extensions', 'numbers', 'operations', 'operations')
-BigDecimal.include Numbers::Operations
+Integer.include Numbers::Operations
 
 
 require Rails.root.join('lib', 'core_extensions', 'dates', 'operations', 'operations')
